LayoutTests/http/tests/workers/worker-importScripts-banned-mimetype.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
startWorker("resources/worker-importScripts-banned-mimetype.py");
</script>
</body>
</html>
